requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
fastapi==0.115.12
h11==0.14.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
pydantic==2.11.3
pydantic_core==2.33.1
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.0

python-dotenv~=1.1.0
Pyrogram~=2.0.106
openai~=1.72.0